+++ index.js
@@ -1,36 +1,7 @@
-var series = require('continuable-series');
-var path = require('path');
-var fs = require('fs');
-var jsonParse = require('safe-json-parse');
-
-var createTasks = require('./create-tasks.js');
-var createNextVersion = require('./create-next-version.js');
-
-function readJson(file, cb) {
-    fs.readFile(file, function (err, buf) {
-        if (err) {
-            return cb(err);
-        }
-
-        jsonParse(String(buf), cb);
-    });
-}
-
-function computeVersion(opts, cb) {
-    var package = path.join(opts.folder, 'package.json');
-
-    readJson(package, function (err, json) {
-        if (err) {
-            return cb(err);
-        }
-
-        var currentVersion = json.version;
-        var nextVersion = createNextVersion(
-            currentVersion, opts);
-
-        cb(null, nextVersion);
-    });
-}
+var bumpMinor = require('./tasks/bump-minor.js');
+var updateChangelog = require('./tasks/update-changelog.js');
+var commitChanges = require('./tasks/commit-changes.js');
+var computeVersion = require('./tasks/compute-next-version.js');
 
 function main(opts, cb) {
     if (typeof opts === 'string') {
@@ -42,12 +13,22 @@ function main(opts, cb) {
             return cb(err);
         }
 
-        series(createTasks({
-            folder: opts.folder,
-            logFlags: opts.logFlags,
-            nextVersion: nextVersion
-        }), function (err) {
-            cb(err, nextVersion);
+        opts.nextVersion = nextVersion;
+
+        bumpMinor(opts, function (err) {
+            if (err) {
+                return cb(err);
+            }
+
+            updateChangelog(opts, function (err) {
+                if (err) {
+                    return cb(err);
+                }
+
+                commitChanges(opts, function (err) {
+                    cb(err, nextVersion);
+                });
+            });
         });
     }
